requirements.txt
allure-python-commons~=2.15.0
pytest~=8.4.0
selenium~=4.34.2